+++ Samples~/Scripts/IssueAndTesting/Issue/Issue389/BuildObject.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue389
+{
+    [Serializable]
+    public class BuildObject
+    {
+        [DefaultExpand]
+        public BuildingAttackBrain buildingAttackBrain;
+
+        private bool InMatchedType => true;
+    }
+}