+++ Samples~/Scripts/IssueAndTesting/Issue/Issue389/NoSuchFunction.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue389
+{
+    [Serializable]
+    public class NoSuchFunction
+    {
+        [DefaultExpand]
+        public BuildingAttackBrain buildingAttackBrain;
+    }
+}